# api/openapi.yaml
openapi: 3.0.4
info:
  title: Favorite Game Gauntlet
  version: 0.9.0
paths:
...
  /games/roll:
    post:
      operationId: makeGameRoll
      tags: [ Games ]
      summary: Select a random game from the wishlist
      description: |
        Select a random game from your unplayed list.<br>
        The selected game will become the current game.<br>
        You will then be able to start the timer.
      responses:
        '200':
          description: A game roll was performed and the selected game returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No unplayed games available for rolling.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Game roll not allowed due to current game state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /wheel-effects/history:
    get:
      operationId: getWheelEffectHistory
      tags: [ WheelEffects ]
      summary: Get a list of all effect rolls
      description: |
        You can get a list of all effect rolls that have been received or rolled,
        as well as all the information about them.
      responses:
        '200':
          description: Effect history retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolledWheelEffects'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /wheel-effects/available:
    get:
      operationId: getAvailableWheelEffects
      tags: [ WheelEffects ]
      summary: Get a list of effects available for rolling
      description: |
        You can get a list of effects that are available to drop using the throw effects command.
      responses:
        '200':
          description: List of available effects successfully retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WheelEffects'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /wheel-effects/available/roll:
    post:
      operationId: makeAvailableWheelEffectRoll
      tags: [ WheelEffects ]
      summary: Make the roll of the effect available
      description: |
        Allows you to obtain a random effect from those that haven't dropped yet.<br>
        A list of available effects can be obtained using another command.
      responses:
        '200':
          description: Effect roll performed successfully and returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolledWheelEffects'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No available effects to roll.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Effect roll not allowed due to current game or timer state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /wheel-effects/available/roll/apply:
    post:
      operationId: applyAvailableWheelEffectRoll
      tags: [ WheelEffects ]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInfos'
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/UserLogin'
                    - $ref: '#/components/schemas/FreePointChange'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No available effects to roll.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Effect roll not allowed due to current game or timer state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /games/unplayed:
    post:
      operationId: addUnplayedGames
      tags: [ Games ]
      deprecated: true
      summary: Add game wishlist
      description: |
        Add the games you want to play here.<br>
        You can use another command to get a random game from this list.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WishlistGames'
      responses:
        '200':
          description: Unplayed games were saved successfully.
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      operationId: getUnplayedGames
      tags: [ Games ]
      deprecated: true
      summary: Get game wishlist
      description: |
        You can see the games you might get here.<br>
        You can use another command to get a random game from this list.
      responses:
        '200':
          description: Unplayed games were retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WishlistGames'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /effects/history:
    get:
      operationId: getEffectHistory
      tags: [ Effects ]
      deprecated: true
      summary: Get a list of all effect rolls
      description: |
        You can get a list of all effect rolls that have been received or rolled,
        as well as all the information about them.
      responses:
        '200':
          description: Effect history retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolledWheelEffects'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /effects/roll:
    post:
      operationId: makeEffectRoll
      tags: [ Effects ]
      deprecated: true
      summary: Make the roll of the effect available
      description: |
        Allows you to obtain a random effect from those that haven't dropped yet.<br>
        A list of available effects can be obtained using another command.
      responses:
        '200':
          description: Effect roll performed successfully and returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolledWheelEffects'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No available effects to roll.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Effect roll not allowed due to current game or timer state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /effects/available:
    get:
      operationId: getAvailableEffects
      tags: [ Effects ]
      deprecated: true
      summary: Get a list of effects available for rolling
      description: |
        You can get a list of effects that are available to drop using the throw effects command.
      responses:
        '200':
          description: List of available effects successfully retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WheelEffects'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /effects/available/count:
    get:
      operationId: getAvailableEffectsCount
      tags: [ Effects ]
      deprecated: true
      summary: Get the number of effects available for rolling
      description: |
        Returns the total number of effects that are currently available for rolling.<br>
        This value represents how many effects can still be obtained via the roll command.
      responses:
        '200':
          description: Number of available effects successfully retrieved.
          content:
            application/json:
              schema:
                type: integer
                format: int
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Available effects data not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
...
  schemas:
...
    UserLogin:
      type: string
      pattern: '^\w+$'
      minLength: 3
      maxLength: 35
    Name:
      type: string
      pattern: '^.+$'
      minLength: 1
      maxLength: 70
...
    Duration:
      type: string
      format: duration
      example: PT2h10m5s
      description: The duration notation as defined by ISO 8601
    Points:
      type: integer
      format: int
      minimum: 0
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required: [ code, message ]
...
    UserStats:
      type: object
      properties:
        territoryPoints:
          $ref: '#/components/schemas/Points'
        freePoints:
          $ref: '#/components/schemas/Points'
        availableRolls:
          $ref: '#/components/schemas/Points'
        territoryHours:
          $ref: '#/components/schemas/Points'
        experiencePoints:
          $ref: '#/components/schemas/Points'
      required: [ territoryPoints, freePoints, availableRolls ]
    UserInfo:
      allOf:
        - type: object
          properties:
            login:
              $ref: '#/components/schemas/UserLogin'
            displayName:
              $ref: '#/components/schemas/Name'
          required: [ displayName ]
        - $ref: '#/components/schemas/UserStats'
    UserInfos:
      type: array
      items:
        $ref: '#/components/schemas/UserInfo'
    WishlistGame:
      type: object
      properties:
        name:
          $ref: '#/components/schemas/Name'
      required: [ name ]
    WishlistGames:
      type: array
      items:
        $ref: '#/components/schemas/WishlistGame'
    Game:
      type: object
      properties:
        name:
          $ref: '#/components/schemas/Name'
        state:
          type: string
          enum: [ 'started', 'finished', 'cancelled' ]
        timeSpent:
          $ref: '#/components/schemas/Duration'
        finishDate:
          type: string
          format: date-time
      required: [ name, state, timeSpent ]
...
    WheelEffect:
      type: object
      properties:
        name:
          $ref: '#/components/schemas/Name'
        description:
          type: string
      required: [ name ]
    WheelEffects:
      type: array
      items:
        $ref: '#/components/schemas/WheelEffect'
    RolledWheelEffect:
      allOf:
        - $ref: '#/components/schemas/WheelEffect'
        - type: object
          properties:
            position:
              type: integer
              format: int
            rollDate:
              type: string
              format: date-time
          required: [ rollDate ]
    RolledWheelEffects:
      type: array
      items:
        $ref: '#/components/schemas/RolledWheelEffect'
    VictoryPointChange:
      type: object
      properties:
        changeSource:
          type: string
        changeValue:
          type: integer
          format: int
        actualChangeValue:
          type: integer
          format: int
        finalValue:
          $ref: '#/components/schemas/Points'
        changeDate:
          type: string
          format: date-time
      required: [ changeSource, changeValue, finalValue, changeDate ]
...
    FreePointChange:
      allOf:
        - $ref: '#/components/schemas/VictoryPointChange'
        - type: object
          properties:
            changeSource:
              enum: [ 'own-wheel-effect', 'other-wheel-effect', 'quest', 'base-teleport', 'sandstorm', 'other' ]
            wheelEffectName:
              $ref: '#/components/schemas/Name'
...